+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-08-29 14:21:34
-// Commit: feat(storage): add bi-directional event bus handler for storage
+// Updated at: 2026-08-31 16:59:21
+// Commit: style(storage): harmonize light and dark theme palette for storage
 
 class storageComponentState {
-  final String commitRef = '4025';
-  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
-  final DateTime timestamp = DateTime.parse('2026-08-29T14:21:34+05:00');
+  final String commitRef = '4086';
+  final String lastMessage = 'style(storage): harmonize light and dark theme palette for storage';
+  final DateTime timestamp = DateTime.parse('2026-08-31T16:59:21+05:00');
 }