+++ CHANGELOG.md
@@ -4083,3 +4083,4 @@
 - [2026-08-31 16:34] perf(theme): apply RepaintBoundary on animated theme widgets
 - [2026-08-31 16:49] fix(forms): fix memory leak by cancelling forms subscription
 - [2026-08-31 16:50] test(theme): implement widget integration tests for theme
+- [2026-08-31 16:59] style(storage): harmonize light and dark theme palette for storage